Move ElementsRegTestParams outside of bitcoinj
